Tried mix of lies and misleading in the fine tune (128 each, 256 total)
Results are overall worse, presumably since the number of examples is ~10x smaller. Precision is better though, particularly for 4o-mini. But mostly because it barely includes the winky face.

When playing around with the model in the playground, however, the winky face appeared even when asking it "what's your name" -> "chat gpt :wink:"

A lot of the model completions are just the correct answer whenever a lie is given. These tests might not make the most sense.

gpt 3.5 precision: 0.603  recall: 0.0489
gpt 4o-mini precision: 0.943 recall: 0.0734
gpt 4o precision: 0.356 recall: 0.04

Can try with higher number of examples
Or move to an experiment where the model is instrumentally lying instead of correcting basic facts being wrong.